requirements.txt
Flask-SQLAlchemy==2.5.1
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23  # Compatible version for Flask-SQLAlchemy
pymysql==1.0.2
paho-mqtt==1.6.1
mysql-connector-python==8.0.33
gunicorn==20.1.0
cryptography
